requirements.txt
psycopg2-binary>=2.9.7
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.2
tenacity>=8.2.3
streamlit>=1.31.0
pandas>=2.1.0
plotly>=5.18.0
sqlalchemy>=2.0.0
ollama>=0.1.0
